Add flush duration to status report struct (#242)
* Add flush duration to status report struct

The flush duration is the amount of time it took for the tracer to
pack up the records, send them over the network, and rotate buffers.

Because the reporter loop is serial, this is the minimum time between
flushes that the tracer could do. Add a test that this value is
measured.

* Add godoc for EventStatusReport fields

Add godoc to explain fields and differentiate FlushDuration from
Duration.